refactor(vendordashboard): split dashboard page into modular components and shared utils; normalize text encoding
