+++ web/app/vendordashboard/constants.ts
@@ -1,0 +1,16 @@
+export const CAT_COLORS: Record<string, string> = {
+  Produce: "#2D6A4F",
+  Dairy: "#1565C0",
+  Meat: "#C1292E",
+  Bakery: "#D4700A",
+  Pantry: "#6D6560",
+};
+
+export const SRC_LABELS: Record<string, string> = {
+  vendor: "Vendor",
+  api: "API",
+  community: "Community",
+  scraper: "Scraped",
+};
+
+export const SALES_PER_PAGE = 9;
