+++ web/app/vendordashboard/page.tsx
@@ -1,7 +1,22 @@
-"use client";
+﻿"use client";
 import React, { useState, useRef, useEffect, useCallback } from "react";
-import VendorMap from "@/components/VendorMap";
 import ProfileDropdown from "@/components/ProfileDropdown";
+import PaginationControls from "./components/PaginationControls";
+import ProductModal, { type NewProductFormValues } from "./components/ProductModal";
+import ReviewsTab from "./components/ReviewsTab";
+import StoreInfoTab from "./components/StoreInfoTab";
+import StoreInfoModal, { type StoreInfoFormValues } from "./components/StoreInfoModal";
+import { SALES_PER_PAGE } from "./constants";
+import {
+  catColor,
+  csvEscape,
+  fmtDate,
+  fmtTime,
+  srcLabel,
+  srcPillClass,
+} from "./lib/formatters";
+import { transformCatalog, transformProducts, transformReviews } from "./lib/transformers";
+import type { CatalogProduct, PriceHistoryEntry, Store, StoreProduct, StoreReview } from "./types";
 import {
   getVendorStore,
   getStoreProducts,
@@ -15,244 +30,6 @@ import {
   deleteStoreProduct,
   updateVendorStoreInfo,
 } from "./actions";
-
-/* ═══ INTERFACES ═══ */
-
-interface StoreProduct {
-  id: string;
-  product_id: string;
-  name: string;
-  brand: string | null;
-  category: string;
-  unit_size: string;
-  price: number;
-  sale_price: number | null;
-  in_stock: boolean;
-  data_source: string;
-  updated_at: string;
-}
-
-interface PriceHistoryEntry {
-  price: number;
-  sale_price: number | null;
-  created_at: string;
-}
-
-interface StoreReview {
-  id: string;
-  rating: number;
-  comment: string | null;
-  user_name: string;
-  created_at: string;
-}
-
-interface Store {
-  id: string;
-  name: string;
-  chain: string | null;
-  address: string | null;
-  zip_code: string | null;
-  lat: number | null;
-  lng: number | null;
-  phone: string | null;
-  website_url: string | null;
-}
-
-interface CatalogProduct {
-  id: string;
-  name: string;
-  brand: string | null;
-  category: string;
-  unit_size: string;
-}
-
-/* ═══ HELPERS ═══ */
-
-const CAT_COLORS: Record<string, string> = {
-  Produce: "#2D6A4F",
-  Dairy: "#1565C0",
-  Meat: "#C1292E",
-  Bakery: "#D4700A",
-  Pantry: "#6D6560",
-};
-
-const SRC_LABELS: Record<string, string> = {
-  vendor: "Vendor",
-  api: "API",
-  community: "Community",
-  scraper: "Scraped",
-};
-
-const catColor = (c: string): string => CAT_COLORS[c] || "#8B8680";
-const srcLabel = (s: string): string => SRC_LABELS[s] || s;
-
-const srcPillClass = (s: string): string => {
-  if (s === "vendor") return "bg-green-100 text-green-800";
-  if (s === "api") return "bg-blue-100 text-blue-700";
-  if (s === "community") return "bg-orange-100 text-orange-700";
-  return "bg-stone-100 text-stone-500";
-};
-
-const fmtTime = (iso: string): string => {
-  const d = new Date(iso);
-  const now = new Date();
-  const m = Math.floor((now.getTime() - d.getTime()) / 60000);
-  if (m < 60) return `${m}m ago`;
-  const h = Math.floor(m / 60);
-  if (h < 24) return `${h}h ago`;
-  return d.toLocaleDateString("en-US", { month: "short", day: "numeric" });
-};
-
-const fmtDate = (iso: string): string =>
-  new Date(iso).toLocaleDateString("en-US", { month: "short", day: "numeric" });
-
-const starsStr = (n: number): string => "★".repeat(n) + "☆".repeat(5 - n);
-
-const csvEscape = (value: string | number | boolean | null | undefined): string => {
-  if (value === null || value === undefined) return "";
-  const str = String(value);
-  if (/[",\r\n]/.test(str)) return `"${str.replace(/"/g, '""')}"`;
-  return str;
-};
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-const transformProducts = (raw: any[]): StoreProduct[] =>
-  raw.map((sp) => ({
-    id: sp.id,
-    product_id: sp.products?.id || "",
-    name: sp.products?.name || "Unknown",
-    brand: sp.products?.brand || null,
-    category: sp.products?.product_categories?.name || "Other",
-    unit_size: sp.products?.unit_size || "",
-    price: sp.price,
-    sale_price: sp.sale_price,
-    in_stock: sp.in_stock,
-    data_source: sp.data_source,
-    updated_at: sp.updated_at,
-  }));
-
-const transformReviews = (raw: any[]): StoreReview[] =>
-  raw.map((r) => ({
-    id: r.id,
-    rating: r.rating,
-    comment: r.comment,
-    user_name: r.users
-      ? `${r.users.first_name || ""} ${r.users.last_name?.charAt(0) || ""}.`.trim()
-      : "Anonymous",
-    created_at: r.created_at,
-  }));
-
-const transformCatalog = (raw: any[]): CatalogProduct[] =>
-  raw.map((p) => ({
-    id: p.id,
-    name: p.name,
-    brand: p.brand,
-    category: p.product_categories?.name || "Other",
-    unit_size: p.unit_size || "",
-  }));
-/* eslint-enable @typescript-eslint/no-explicit-any */
-
-/* ═══ PAGINATION CONTROLS ═══ */
-
-const PaginationControls: React.FC<{
-  page: number;
-  totalPages: number;
-  onChange: (p: number) => void;
-}> = ({ page, totalPages, onChange }) => {
-  const [jumpValue, setJumpValue] = useState("");
-
-  const go = (p: number) => onChange(Math.max(1, Math.min(totalPages, p)));
-
-  const handleJump = () => {
-    const n = parseInt(jumpValue, 10);
-    if (!isNaN(n)) go(n);
-    setJumpValue("");
-  };
-
-  // Build the list of page numbers to render, collapsing long ranges with "…"
-  const pages: (number | "…")[] = [];
-  if (totalPages <= 7) {
-    for (let i = 1; i <= totalPages; i++) pages.push(i);
-  } else {
-    pages.push(1);
-    if (page > 3) pages.push("…");
-    const start = Math.max(2, page - 1);
-    const end = Math.min(totalPages - 1, page + 1);
-    for (let i = start; i <= end; i++) pages.push(i);
-    if (page < totalPages - 2) pages.push("…");
-    pages.push(totalPages);
-  }
-
-  const btnBase =
-    "h-8 min-w-[32px] px-2 rounded-lg text-xs font-semibold transition-colors cursor-pointer flex items-center justify-center";
-  const btnIdle = "border border-stone-200 bg-white text-stone-600 hover:bg-stone-50";
-  const btnActive = "bg-green-800 text-white border border-green-800";
-  const btnDisabled = "opacity-40 cursor-not-allowed";
-
-  return (
-    <div className="flex items-center gap-1.5 flex-wrap justify-end">
-      <button
-        onClick={() => go(page - 1)}
-        disabled={page === 1}
-        className={`${btnBase} ${btnIdle} ${page === 1 ? btnDisabled : ""}`}
-      >
-        ← Prev
-      </button>
-
-      {pages.map((p, i) =>
-        p === "…" ? (
-          <span key={`e${i}`} className="px-1 text-stone-400 text-xs select-none">
-            …
-          </span>
-        ) : (
-          <button
-            key={p}
-            onClick={() => go(p)}
-            className={`${btnBase} ${p === page ? btnActive : btnIdle}`}
-          >
-            {p}
-          </button>
-        ),
-      )}
-
-      <button
-        onClick={() => go(page + 1)}
-        disabled={page === totalPages}
-        className={`${btnBase} ${btnIdle} ${page === totalPages ? btnDisabled : ""}`}
-      >
-        Next
-      </button>
-
-      <div className="flex items-center gap-1.5 ml-3 pl-3 border-l border-stone-200">
-        <span className="text-xs text-stone-400">Go to</span>
-        <input
-          type="number"
-          min={1}
-          max={totalPages}
-          value={jumpValue}
-          onChange={(e) => setJumpValue(e.target.value)}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") {
-              e.preventDefault();
-              handleJump();
-            }
-          }}
-          placeholder={String(page)}
-          className="w-14 px-2 py-1.5 text-xs font-semibold text-center border border-stone-200 rounded-lg outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 bg-white"
-        />
-        <button
-          onClick={handleJump}
-          disabled={!jumpValue}
-          className={`${btnBase} ${btnIdle} ${!jumpValue ? btnDisabled : ""}`}
-        >
-          Go
-        </button>
-      </div>
-    </div>
-  );
-};
-
-/* ═══ COMPONENT ═══ */
 
 const VendorDashboard: React.FC = () => {
   /* ── Data state ── */
@@ -271,7 +48,7 @@ const VendorDashboard: React.FC = () => {
   const [pendingDeleteId, setPendingDeleteId] = useState<string | null>(null);
   const [deletingId, setDeletingId] = useState<string | null>(null);
   const [showStoreModal, setShowStoreModal] = useState(false);
-  const [storeForm, setStoreForm] = useState({
+  const [storeForm, setStoreForm] = useState<StoreInfoFormValues>({
     name: "",
     chain: "",
     address: "",
@@ -290,7 +67,7 @@ const VendorDashboard: React.FC = () => {
   const [catalogPrice, setCatalogPrice] = useState("");
   const [catalogSalePrice, setCatalogSalePrice] = useState("");
   const [catalogInStock, setCatalogInStock] = useState(true);
-  const [newProduct, setNewProduct] = useState({
+  const [newProduct, setNewProduct] = useState<NewProductFormValues>({
     name: "",
     brand: "",
     category: "Produce",
@@ -305,7 +82,6 @@ const VendorDashboard: React.FC = () => {
   const searchInputRef = useRef<HTMLInputElement>(null);
 
   /* ── Pagination state ── */
-  const SALES_PER_PAGE = 9;
   const [productsPerPage, setProductsPerPage] = useState(10);
   const [productPage, setProductPage] = useState(1);
   const [salePage, setSalePage] = useState(1);
@@ -1219,7 +995,7 @@ const VendorDashboard: React.FC = () => {
           </div>
         </div>
 
-        {/* ── On Sale — col-span-3 ── */}
+        {/* On Sale */}
         <div className="bg-white rounded-2xl p-7 border border-black/[0.05] hover:shadow-xl hover:-translate-y-0.5 transition-all duration-300 col-span-3">
           <div className="flex items-center gap-2 text-[11px] font-semibold uppercase tracking-widest text-stone-400 mb-5">
             <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
@@ -1280,692 +1056,63 @@ const VendorDashboard: React.FC = () => {
       </div>
       )}
 
-      {/* ═══════════════════════════════════════════ */}
-      {/* ══ REVIEWS TAB ══ */}
-      {/* ═══════════════════════════════════════════ */}
+    
+      {/* REVIEWS TAB */}
+  
       {activeTab === "reviews" && (
-      <div className="max-w-[1320px] mx-auto px-8 pb-12 grid grid-cols-3 gap-5">
-
-        {/* ── Overall Rating Card ── */}
-        <div className="bg-white rounded-2xl p-7 border border-black/[0.05] hover:shadow-xl hover:-translate-y-0.5 transition-all duration-300">
-          <div className="flex items-center gap-2 text-[11px] font-semibold uppercase tracking-widest text-stone-400 mb-6">
-            <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <polygon points="12 2 15.09 8.26 22 9.27 17 14.14 18.18 21.02 12 17.77 5.82 21.02 7 14.14 2 9.27 8.91 8.26 12 2"/>
-            </svg>
-            Overall Rating
-          </div>
-
-          <div className="text-center mb-8">
-            <div className="fraunces text-[72px] font-semibold leading-none text-orange-600">{avgRating}</div>
-            <div className="text-[26px] text-orange-500 tracking-widest mt-2">
-              {"★".repeat(Math.round(parseFloat(avgRating)))}{"☆".repeat(5 - Math.round(parseFloat(avgRating)))}
-            </div>
-            <div className="text-sm text-stone-400 mt-2">Based on {reviews.length} reviews</div>
-          </div>
-
-          <div className="flex flex-col gap-3">
-            {ratingCounts.map(({ rating, count, pct }) => (
-              <div key={rating} className="flex items-center gap-3">
-                <span className="text-sm font-medium text-stone-500 w-4 text-right">{rating}</span>
-                <svg width="14" height="14" viewBox="0 0 24 24" fill="#D4700A" stroke="none">
-                  <polygon points="12 2 15.09 8.26 22 9.27 17 14.14 18.18 21.02 12 17.77 5.82 21.02 7 14.14 2 9.27 8.91 8.26 12 2"/>
-                </svg>
-                <div className="flex-1 bg-stone-100 rounded-full h-2.5 overflow-hidden">
-                  <div className="h-full rounded-full bg-orange-400 transition-all duration-500" style={{ width: `${pct}%` }}/>
-                </div>
-                <span className="text-sm font-medium text-stone-500 w-6 text-right">{count}</span>
-              </div>
-            ))}
-          </div>
-
-          <div className="mt-8 pt-6 border-t border-stone-100 grid grid-cols-2 gap-4">
-            <div className="bg-stone-50 rounded-xl px-4 py-3 border border-stone-100 text-center">
-              <div className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1">5-Star</div>
-              <div className="fraunces text-xl font-semibold text-green-800">
-                {reviews.length ? Math.round((ratingCounts[0].count / reviews.length) * 100) : 0}%
-              </div>
-            </div>
-            <div className="bg-stone-50 rounded-xl px-4 py-3 border border-stone-100 text-center">
-              <div className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1">Avg Score</div>
-              <div className="fraunces text-xl font-semibold text-orange-600">{avgRating}</div>
-            </div>
-          </div>
-        </div>
-
-        {/* ── Customer Reviews List ── */}
-        <div className="bg-white rounded-2xl p-7 border border-black/[0.05] hover:shadow-xl hover:-translate-y-0.5 transition-all duration-300 col-span-2">
-          <div className="flex justify-between items-center mb-5">
-            <div className="flex items-center gap-2 text-[11px] font-semibold uppercase tracking-widest text-stone-400">
-              <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z"/>
-              </svg>
-              Customer Reviews
-            </div>
-            <span className="text-xs font-medium text-stone-400 bg-stone-100 px-3 py-1.5 rounded-full">
-              {reviews.length} total
-            </span>
-          </div>
-
-          {reviews.length === 0 ? (
-            <div className="text-center py-12 text-stone-400 text-sm">No reviews yet.</div>
-          ) : (
-            <div className="flex flex-col">
-              {reviews.map((r, i) => (
-                <div key={r.id} className={`py-5 ${i < reviews.length - 1 ? "border-b border-stone-100" : ""}`}>
-                  <div className="flex justify-between items-start mb-2.5">
-                    <div className="flex items-center gap-3">
-                      <div
-                        className="w-10 h-10 rounded-xl flex items-center justify-center text-sm font-bold text-white"
-                        style={{
-                          background: `linear-gradient(135deg, ${
-                            ["#2D6A4F", "#D94F30", "#1565C0", "#D4700A", "#6D6560"][i % 5]
-                          }, ${
-                            ["#52B788", "#F4A261", "#42A5F5", "#FFB74D", "#A1887F"][i % 5]
-                          })`,
-                        }}
-                      >
-                        {r.user_name.charAt(0)}
-                      </div>
-                      <div>
-                        <div className="font-semibold text-[15px]">{r.user_name}</div>
-                        <div className="text-orange-500 text-sm tracking-wider mt-0.5">{starsStr(r.rating)}</div>
-                      </div>
-                    </div>
-                    <span className="text-xs text-stone-400 bg-stone-50 px-3 py-1.5 rounded-full border border-stone-100">
-                      {fmtDate(r.created_at)}
-                    </span>
-                  </div>
-                  {r.comment ? (
-                    <div className="text-sm text-stone-600 leading-relaxed ml-[52px] bg-stone-50 rounded-xl px-4 py-3 border border-stone-100">
-                      &ldquo;{r.comment}&rdquo;
-                    </div>
-                  ) : (
-                    <div className="text-sm text-stone-300 italic ml-[52px]">No written review</div>
-                  )}
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-
-      </div>
+        <ReviewsTab avgRating={avgRating} reviews={reviews} ratingCounts={ratingCounts} />
       )}
 
-      {/* ══ STORE INFO TAB ══ */}
+      {/* STORE INFO TAB */}
       {activeTab === "store info" && (
-      <div className="max-w-[1320px] mx-auto px-8 pb-12 grid grid-cols-5 gap-5 items-start">
-
-        <div className="bg-white rounded-2xl p-7 border border-black/[0.05] hover:shadow-xl hover:-translate-y-0.5 transition-all duration-300 col-span-2">
-          <div className="flex justify-between items-center mb-5">
-            <div className="flex items-center gap-2 text-[11px] font-semibold uppercase tracking-widest text-stone-400">
-              <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                <path d="M3 9l9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z"/>
-                <polyline points="9 22 9 12 15 12 15 22"/>
-              </svg>
-              Store Details
-            </div>
-            <button
-              onClick={openStoreInfoModal}
-              className="border border-green-800 text-green-800 bg-transparent px-3.5 py-1.5 rounded-xl text-xs font-semibold cursor-pointer hover:bg-green-50 transition-colors"
-            >
-              Edit Store Info
-            </button>
-          </div>
-
-          <div className="flex items-center gap-4 mb-6">
-            <div
-              className="w-14 h-14 rounded-2xl flex items-center justify-center text-white font-bold text-xl fraunces"
-              style={{ background: "linear-gradient(135deg,#D94F30,#F4A261)" }}
-            >{storeInitials}</div>
-            <div>
-              <div className="fraunces text-[22px] font-semibold tracking-tight">{storeName}</div>
-              <div className="text-sm text-stone-400">{store?.address || "—"}</div>
-            </div>
-          </div>
-
-          <div className="grid grid-cols-1 gap-4">
-            {([
-              {
-                label: "Store Name",
-                value: store?.name,
-                icon: (
-                  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                    <path d="M3 9l9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z"/>
-                    <polyline points="9 22 9 12 15 12 15 22"/>
-                  </svg>
-                ),
-              },
-              {
-                label: "Address",
-                value: store?.address,
-                icon: (
-                  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                    <path d="M21 10c0 7-9 13-9 13s-9-6-9-13a9 9 0 0 1 18 0z"/>
-                    <circle cx="12" cy="10" r="3"/>
-                  </svg>
-                ),
-              },
-              {
-                label: "Phone",
-                value: store?.phone,
-                icon: (
-                  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                    <path d="M22 16.92v3a2 2 0 0 1-2.18 2 19.79 19.79 0 0 1-8.63-3.07 19.5 19.5 0 0 1-6-6 19.79 19.79 0 0 1-3.07-8.67A2 2 0 0 1 4.11 2h3a2 2 0 0 1 2 1.72 12.84 12.84 0 0 0 .7 2.81 2 2 0 0 1-.45 2.11L8.09 9.91a16 16 0 0 0 6 6l1.27-1.27a2 2 0 0 1 2.11-.45 12.84 12.84 0 0 0 2.81.7A2 2 0 0 1 22 16.92z"/>
-                  </svg>
-                ),
-              },
-              {
-                label: "Website",
-                value: store?.website_url,
-                icon: (
-                  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                    <circle cx="12" cy="12" r="10"/>
-                    <line x1="2" y1="12" x2="22" y2="12"/>
-                    <path d="M12 2a15.3 15.3 0 0 1 4 10 15.3 15.3 0 0 1-4 10 15.3 15.3 0 0 1-4-10 15.3 15.3 0 0 1 4-10z"/>
-                  </svg>
-                ),
-              },
-              {
-                label: "Zip Code",
-                value: store?.zip_code,
-                icon: (
-                  <svg width="15" height="15" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                    <path d="M4 4h16c1.1 0 2 .9 2 2v12c0 1.1-.9 2-2 2H4c-1.1 0-2-.9-2-2V6c0-1.1.9-2 2-2z"/>
-                    <polyline points="22,6 12,13 2,6"/>
-                  </svg>
-                ),
-              },
-            ] as { label: string; value: string | null | undefined; icon: React.ReactNode }[]).map((f, i) => (
-              <div key={i} className="bg-stone-50 rounded-xl px-4 py-3.5 border border-stone-100 flex items-start gap-3">
-                <div className="text-stone-400 mt-0.5 flex-shrink-0">{f.icon}</div>
-                <div>
-                  <div className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1">{f.label}</div>
-                  <div className="text-sm font-medium break-words">{f.value || "—"}</div>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        {/* ── Map ── */}
-        <div className="bg-white rounded-2xl border border-black/[0.05] hover:shadow-xl hover:-translate-y-0.5 transition-all duration-300 overflow-hidden col-span-3">
-          <div className="flex items-center gap-2 text-[11px] font-semibold uppercase tracking-widest text-stone-400 px-7 pt-7 pb-4">
-            <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <path d="M21 10c0 7-9 13-9 13s-9-6-9-13a9 9 0 0 1 18 0z"/>
-              <circle cx="12" cy="10" r="3"/>
-            </svg>
-            Location
-          </div>
-          <div style={{ height: "620px" }}>
-            <VendorMap
-              address={store?.address || ""}
-              lat={store?.lat || 40.6892}
-              lng={store?.lng || -73.9857}
-            />
-          </div>
-        </div>
-
-      </div>
+        <StoreInfoTab
+          store={store}
+          storeInitials={storeInitials}
+          storeName={storeName}
+          onEditStoreInfo={openStoreInfoModal}
+        />
       )}
 
       {showStoreModal && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center" onClick={closeStoreInfoModal}>
-          <div className="absolute inset-0 bg-black/40 backdrop-blur-sm" />
-
-          <div
-            className="relative bg-white rounded-2xl w-full max-w-[620px] max-h-[85vh] overflow-hidden flex flex-col"
-            style={{ boxShadow: "0 25px 60px rgba(0,0,0,0.15)" }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            <div className="px-7 pt-6 pb-4 border-b border-stone-100">
-              <div className="flex justify-between items-center">
-                <div className="fraunces text-xl font-semibold tracking-tight">Edit Store Info</div>
-                <button
-                  onClick={closeStoreInfoModal}
-                  disabled={savingStore}
-                  className="w-8 h-8 rounded-lg flex items-center justify-center text-stone-400 hover:bg-stone-100 hover:text-stone-600 transition-colors border-none bg-transparent cursor-pointer disabled:opacity-40 disabled:cursor-not-allowed"
-                >
-                  <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                    <line x1="18" y1="6" x2="6" y2="18"/><line x1="6" y1="6" x2="18" y2="18"/>
-                  </svg>
-                </button>
-              </div>
-            </div>
-
-            <form
-              onSubmit={(e) => {
-                e.preventDefault();
-                saveStoreInfo();
-              }}
-              className="px-7 py-5 overflow-y-auto flex-1"
-            >
-              {storeFormError && (
-                <div className="mb-4 bg-red-50 border border-red-200 text-red-700 text-sm rounded-xl px-4 py-3">
-                  {storeFormError}
-                </div>
-              )}
-
-              <div className="flex flex-col gap-4">
-                <div>
-                  <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Store Name *</label>
-                  <input
-                    value={storeForm.name}
-                    onChange={(e) => setStoreForm((prev) => ({ ...prev, name: e.target.value }))}
-                    placeholder="Store name"
-                    className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                  />
-                </div>
-
-                <div className="grid grid-cols-2 gap-4">
-                  <div>
-                    <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Chain</label>
-                    <input
-                      value={storeForm.chain}
-                      onChange={(e) => setStoreForm((prev) => ({ ...prev, chain: e.target.value }))}
-                      placeholder="Optional"
-                      className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                    />
-                  </div>
-                  <div>
-                    <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Zip Code</label>
-                    <input
-                      value={storeForm.zip_code}
-                      onChange={(e) => setStoreForm((prev) => ({ ...prev, zip_code: e.target.value }))}
-                      placeholder="Optional"
-                      className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                    />
-                  </div>
-                </div>
-
-                <div>
-                  <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Address</label>
-                  <input
-                    value={storeForm.address}
-                    onChange={(e) => setStoreForm((prev) => ({ ...prev, address: e.target.value }))}
-                    placeholder="Optional"
-                    className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                  />
-                </div>
-
-                <div className="grid grid-cols-2 gap-4">
-                  <div>
-                    <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Phone</label>
-                    <input
-                      value={storeForm.phone}
-                      onChange={(e) => setStoreForm((prev) => ({ ...prev, phone: e.target.value }))}
-                      placeholder="Optional"
-                      className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                    />
-                  </div>
-                  <div>
-                    <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Website URL</label>
-                    <input
-                      value={storeForm.website_url}
-                      onChange={(e) => setStoreForm((prev) => ({ ...prev, website_url: e.target.value }))}
-                      placeholder="https://example.com"
-                      className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                    />
-                  </div>
-                </div>
-              </div>
-
-              <div className="mt-6 pt-4 border-t border-stone-100 flex justify-end gap-2">
-                <button
-                  type="button"
-                  onClick={closeStoreInfoModal}
-                  disabled={savingStore}
-                  className="border border-stone-200 text-stone-500 bg-transparent px-5 py-2.5 rounded-xl text-sm font-semibold cursor-pointer hover:bg-stone-50 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                >
-                  Cancel
-                </button>
-                <button
-                  type="submit"
-                  disabled={savingStore}
-                  className="bg-green-800 text-white border-none px-5 py-2.5 rounded-xl text-sm font-semibold cursor-pointer hover:bg-green-900 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                >
-                  {savingStore ? "Saving..." : "Save Changes"}
-                </button>
-              </div>
-            </form>
-          </div>
-        </div>
+        <StoreInfoModal
+          storeForm={storeForm}
+          storeFormError={storeFormError}
+          savingStore={savingStore}
+          onClose={closeStoreInfoModal}
+          onSave={saveStoreInfo}
+          onChangeField={(field, value) =>
+            setStoreForm((prev) => ({ ...prev, [field]: value }))
+          }
+        />
       )}
 
-      {/* ══ PRODUCT MODAL ══ */}
+      {/* PRODUCT MODAL */}
       {showProductModal && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center" onClick={closeModal}>
-          <div className="absolute inset-0 bg-black/40 backdrop-blur-sm" />
-
-          <div
-            className="relative bg-white rounded-2xl w-full max-w-[640px] max-h-[85vh] overflow-hidden flex flex-col"
-            style={{ boxShadow: "0 25px 60px rgba(0,0,0,0.15)" }}
-            onClick={(e) => e.stopPropagation()}
-          >
-            {/* Modal Header */}
-            <div className="px-7 pt-6 pb-4 border-b border-stone-100">
-              <div className="flex justify-between items-center mb-4">
-                <div className="fraunces text-xl font-semibold tracking-tight">Add Product</div>
-                <button
-                  onClick={closeModal}
-                  className="w-8 h-8 rounded-lg flex items-center justify-center text-stone-400 hover:bg-stone-100 hover:text-stone-600 transition-colors border-none bg-transparent cursor-pointer"
-                >
-                  <svg width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                    <line x1="18" y1="6" x2="6" y2="18"/><line x1="6" y1="6" x2="18" y2="18"/>
-                  </svg>
-                </button>
-              </div>
-
-              <div className="flex gap-1 bg-stone-100 rounded-full p-1">
-                <button
-                  onClick={() => {
-                    setModalMode("search");
-                    setSelectedCatalogItem(null);
-                    setCatalogPrice("");
-                    setCatalogSalePrice("");
-                    setCatalogInStock(true);
-                    setModalError(null);
-                  }}
-                  className={`flex-1 py-2 rounded-full text-sm font-medium transition-all duration-200 border-none cursor-pointer
-                    ${modalMode === "search"
-                      ? "bg-green-800 text-white shadow-sm"
-                      : "bg-transparent text-stone-500 hover:text-stone-700"}`}
-                >
-                  Search Catalog
-                </button>
-                <button
-                  onClick={() => { setModalMode("create"); setModalError(null); }}
-                  className={`flex-1 py-2 rounded-full text-sm font-medium transition-all duration-200 border-none cursor-pointer
-                    ${modalMode === "create"
-                      ? "bg-green-800 text-white shadow-sm"
-                      : "bg-transparent text-stone-500 hover:text-stone-700"}`}
-                >
-                  Create New
-                </button>
-              </div>
-            </div>
-
-            {/* Modal Body */}
-            <div className="px-7 py-5 overflow-y-auto flex-1">
-
-              {modalError && (
-                <div className="mb-4 bg-red-50 border border-red-200 text-red-700 text-sm rounded-xl px-4 py-3 flex items-start gap-2">
-                  <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" className="flex-shrink-0 mt-0.5">
-                    <circle cx="12" cy="12" r="10"/>
-                    <line x1="12" y1="8" x2="12" y2="12"/>
-                    <line x1="12" y1="16" x2="12.01" y2="16"/>
-                  </svg>
-                  <span>{modalError}</span>
-                </div>
-              )}
-
-              {modalMode === "search" && (
-                <div>
-                  <div className="relative mb-4">
-                    <svg className="absolute left-3 top-1/2 -translate-y-1/2 text-stone-400" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                      <circle cx="11" cy="11" r="8"/><line x1="21" y1="21" x2="16.65" y2="16.65"/>
-                    </svg>
-                    <input
-                      ref={searchInputRef}
-                      value={catalogSearch}
-                      onChange={(e) => { setCatalogSearch(e.target.value); setSelectedCatalogItem(null); setCatalogPrice(""); }}
-                      placeholder="Search by name, brand, or category..."
-                      className="w-full border border-stone-200 rounded-xl pl-10 pr-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                    />
-                  </div>
-
-                  {catalogSearch.trim().length === 0 ? (
-                    <div className="text-center py-10">
-                      <div className="text-stone-300 mb-2">
-                        <svg className="mx-auto" width="40" height="40" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="1.5">
-                          <circle cx="11" cy="11" r="8"/><line x1="21" y1="21" x2="16.65" y2="16.65"/>
-                        </svg>
-                      </div>
-                      <div className="text-sm text-stone-400">Type to search the product catalog</div>
-                      <div className="text-xs text-stone-300 mt-1">Or switch to &ldquo;Create New&rdquo; to add a custom product</div>
-                    </div>
-                  ) : catalogLoading ? (
-                    <div className="text-center py-10">
-                      <div className="inline-flex items-center justify-center w-10 h-10 rounded-full border-2 border-stone-200 border-t-green-700 animate-spin mb-2" />
-                      <div className="text-sm text-stone-400">Searching catalog...</div>
-                    </div>
-                  ) : catalogResults.length === 0 ? (
-                    <div className="text-center py-10">
-                      <div className="text-sm text-stone-400 mb-3">No products found for &ldquo;{catalogSearch}&rdquo;</div>
-                      <button
-                        onClick={() => {
-                          setModalMode("create");
-                          setNewProduct((prev) => ({ ...prev, name: catalogSearch }));
-                        }}
-                        className="bg-green-800 text-white border-none px-4 py-2 rounded-xl text-sm font-semibold cursor-pointer hover:bg-green-900 transition-colors"
-                      >
-                        Create &ldquo;{catalogSearch}&rdquo; as new product
-                      </button>
-                    </div>
-                  ) : (
-                    <div className="flex flex-col gap-2">
-                      {catalogResults.map((item) => {
-                        const isSelected = selectedCatalogItem?.id === item.id;
-                        return (
-                          <div key={item.id}>
-                            <div
-                              className={`flex items-center justify-between p-4 rounded-xl border transition-all cursor-pointer group
-                                ${isSelected
-                                  ? "bg-green-50 border-green-200"
-                                  : "bg-stone-50 border-stone-100 hover:border-green-200 hover:bg-green-50/30"}`}
-                              onClick={() => {
-                                setSelectedCatalogItem(item);
-                                setCatalogPrice("");
-                                setCatalogSalePrice("");
-                                setCatalogInStock(true);
-                                setModalError(null);
-                              }}
-                            >
-                              <div className="flex items-center gap-3">
-                                <div className="w-2.5 h-2.5 rounded-full flex-shrink-0" style={{ background: catColor(item.category) }}/>
-                                <div>
-                                  <div className="font-medium text-sm">{item.name}</div>
-                                  <div className="text-[11px] text-stone-400">
-                                    {item.brand || "Generic"} · {item.category}{item.unit_size ? ` · per ${item.unit_size}` : ""}
-                                  </div>
-                                </div>
-                              </div>
-                              {!isSelected && (
-                                <div className="w-8 h-8 rounded-lg bg-green-100 text-green-800 flex items-center justify-center opacity-0 group-hover:opacity-100 transition-opacity">
-                                  <svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                                    <line x1="12" y1="5" x2="12" y2="19"/><line x1="5" y1="12" x2="19" y2="12"/>
-                                  </svg>
-                                </div>
-                              )}
-                            </div>
-                            {isSelected && (
-                              <div className="mt-2 ml-4 p-4 rounded-xl bg-stone-50 border border-stone-100 flex flex-col gap-3">
-                                <div className="flex items-center gap-3 flex-wrap">
-                                  <div className="flex flex-col gap-1">
-                                    <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider">Price *</label>
-                                    <div className="relative">
-                                      <span className="absolute left-3 top-1/2 -translate-y-1/2 text-stone-400 text-sm">$</span>
-                                      <input
-                                        value={catalogPrice}
-                                        onChange={(e) => setCatalogPrice(e.target.value)}
-                                        placeholder="0.00"
-                                        autoFocus
-                                        inputMode="decimal"
-                                        onKeyDown={(e) => { if (e.key === "Enter") addFromCatalog(); }}
-                                        className="w-28 border border-stone-200 rounded-xl pl-7 pr-3 py-2 text-sm bg-white outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                                      />
-                                    </div>
-                                  </div>
-                                  <div className="flex flex-col gap-1">
-                                    <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider">Sale Price</label>
-                                    <div className="relative">
-                                      <span className="absolute left-3 top-1/2 -translate-y-1/2 text-stone-400 text-sm">$</span>
-                                      <input
-                                        value={catalogSalePrice}
-                                        onChange={(e) => setCatalogSalePrice(e.target.value)}
-                                        placeholder="optional"
-                                        inputMode="decimal"
-                                        onKeyDown={(e) => { if (e.key === "Enter") addFromCatalog(); }}
-                                        className="w-28 border border-stone-200 rounded-xl pl-7 pr-3 py-2 text-sm bg-white outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                                      />
-                                    </div>
-                                  </div>
-                                  <label className="flex items-center gap-2 text-sm text-stone-600 cursor-pointer select-none mt-5">
-                                    <input
-                                      type="checkbox"
-                                      checked={catalogInStock}
-                                      onChange={(e) => setCatalogInStock(e.target.checked)}
-                                      className="w-4 h-4 accent-green-700 cursor-pointer"
-                                    />
-                                    In stock
-                                  </label>
-                                </div>
-                                <div className="flex justify-end">
-                                  <button
-                                    onClick={addFromCatalog}
-                                    disabled={submitting || !catalogPrice || isNaN(parseFloat(catalogPrice)) || parseFloat(catalogPrice) <= 0}
-                                    className="bg-green-800 text-white border-none px-4 py-2 rounded-xl text-sm font-semibold cursor-pointer hover:bg-green-900 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                                  >
-                                    {submitting ? "Adding..." : "Add"}
-                                  </button>
-                                </div>
-                              </div>
-                            )}
-                          </div>
-                        );
-                      })}
-                    </div>
-                  )}
-                </div>
-              )}
-
-              {modalMode === "create" && (
-                <div className="flex flex-col gap-4">
-                  <div>
-                    <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Product Name *</label>
-                    <input
-                      value={newProduct.name}
-                      onChange={(e) => setNewProduct({ ...newProduct, name: e.target.value })}
-                      placeholder="e.g. Organic Whole Milk"
-                      className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                    />
-                  </div>
-
-                  <div className="grid grid-cols-2 gap-4">
-                    <div>
-                      <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Brand</label>
-                      <input
-                        value={newProduct.brand}
-                        onChange={(e) => setNewProduct({ ...newProduct, brand: e.target.value })}
-                        placeholder="e.g. Horizon"
-                        className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                      />
-                    </div>
-                    <div>
-                      <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Category *</label>
-                      <select
-                        value={newProduct.category}
-                        onChange={(e) => setNewProduct({ ...newProduct, category: e.target.value })}
-                        className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all cursor-pointer appearance-none"
-                      >
-                        {Object.keys(CAT_COLORS).map((cat) => (
-                          <option key={cat} value={cat}>{cat}</option>
-                        ))}
-                      </select>
-                    </div>
-                  </div>
-
-                  <div>
-                    <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Unit Size</label>
-                    <input
-                      value={newProduct.unit_size}
-                      onChange={(e) => setNewProduct({ ...newProduct, unit_size: e.target.value })}
-                      placeholder="e.g. lb, gal, dozen"
-                      className="w-full border border-stone-200 rounded-xl px-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                    />
-                  </div>
-
-                  <div className="grid grid-cols-2 gap-4">
-                    <div>
-                      <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Price *</label>
-                      <div className="relative">
-                        <span className="absolute left-4 top-1/2 -translate-y-1/2 text-stone-400 text-sm">$</span>
-                        <input
-                          value={newProduct.price}
-                          onChange={(e) => setNewProduct({ ...newProduct, price: e.target.value })}
-                          placeholder="0.00"
-                          inputMode="decimal"
-                          onKeyDown={(e) => { if (e.key === "Enter") createNewProduct(); }}
-                          className="w-full border border-stone-200 rounded-xl pl-8 pr-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                        />
-                      </div>
-                    </div>
-                    <div>
-                      <label className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-1.5 block">Sale Price</label>
-                      <div className="relative">
-                        <span className="absolute left-4 top-1/2 -translate-y-1/2 text-stone-400 text-sm">$</span>
-                        <input
-                          value={newProduct.sale_price}
-                          onChange={(e) => setNewProduct({ ...newProduct, sale_price: e.target.value })}
-                          placeholder="optional"
-                          inputMode="decimal"
-                          onKeyDown={(e) => { if (e.key === "Enter") createNewProduct(); }}
-                          className="w-full border border-stone-200 rounded-xl pl-8 pr-4 py-3 text-sm bg-stone-50 outline-none focus:border-green-700 focus:ring-2 focus:ring-green-700/10 transition-all"
-                        />
-                      </div>
-                    </div>
-                  </div>
-
-                  <label className="flex items-center gap-2 text-sm text-stone-600 cursor-pointer select-none">
-                    <input
-                      type="checkbox"
-                      checked={newProduct.in_stock}
-                      onChange={(e) => setNewProduct({ ...newProduct, in_stock: e.target.checked })}
-                      className="w-4 h-4 accent-green-700 cursor-pointer"
-                    />
-                    In stock
-                  </label>
-
-                  {newProduct.name.trim() && (
-                    <div className="bg-stone-50 rounded-xl p-4 border border-stone-100 mt-2">
-                      <div className="text-[11px] font-semibold text-stone-400 uppercase tracking-wider mb-2">Preview</div>
-                      <div className="flex items-center gap-2.5">
-                        <div className="w-2.5 h-2.5 rounded-full" style={{ background: catColor(newProduct.category) }}/>
-                        <div>
-                          <div className="font-medium text-sm">{newProduct.name}</div>
-                          <div className="text-[11px] text-stone-400">
-                            {newProduct.brand || "Generic"} · {newProduct.category}
-                            {newProduct.unit_size && ` · per ${newProduct.unit_size}`}
-                            {newProduct.price && ` · $${parseFloat(newProduct.price).toFixed(2)}`}
-                            {newProduct.sale_price && ` · sale $${parseFloat(newProduct.sale_price).toFixed(2)}`}
-                            {!newProduct.in_stock && ` · out of stock`}
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              )}
-            </div>
-
-            {modalMode === "create" && (
-              <div className="px-7 py-4 border-t border-stone-100 flex justify-end gap-2">
-                <button
-                  onClick={closeModal}
-                  className="border border-stone-200 text-stone-500 bg-transparent px-5 py-2.5 rounded-xl text-sm font-semibold cursor-pointer hover:bg-stone-50 transition-colors"
-                >
-                  Cancel
-                </button>
-                <button
-                  onClick={createNewProduct}
-                  disabled={submitting || !newProduct.name.trim() || !newProduct.price}
-                  className="bg-green-800 text-white border-none px-5 py-2.5 rounded-xl text-sm font-semibold cursor-pointer hover:bg-green-900 transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                >
-                  {submitting ? "Adding..." : "Add Product"}
-                </button>
-              </div>
-            )}
-          </div>
-        </div>
+        <ProductModal
+          closeModal={closeModal}
+          modalMode={modalMode}
+          setModalMode={setModalMode}
+          modalError={modalError}
+          setModalError={setModalError}
+          submitting={submitting}
+          searchInputRef={searchInputRef}
+          catalogSearch={catalogSearch}
+          setCatalogSearch={setCatalogSearch}
+          catalogResults={catalogResults}
+          catalogLoading={catalogLoading}
+          selectedCatalogItem={selectedCatalogItem}
+          setSelectedCatalogItem={setSelectedCatalogItem}
+          catalogPrice={catalogPrice}
+          setCatalogPrice={setCatalogPrice}
+          catalogSalePrice={catalogSalePrice}
+          setCatalogSalePrice={setCatalogSalePrice}
+          catalogInStock={catalogInStock}
+          setCatalogInStock={setCatalogInStock}
+          addFromCatalog={addFromCatalog}
+          newProduct={newProduct}
+          setNewProduct={setNewProduct}
+          createNewProduct={createNewProduct}
+        />
       )}
 
     </div>
